fixes get_absolute_path() to work without hydra
