+++ examples/configure_hydra/job_name/with_config_file_override.py
@@ -7,7 +7,9 @@ from hydra.core.hydra_config import HydraConfig
 
 @hydra.main(config_path="config.yaml")
 def experiment(_cfg: DictConfig) -> None:
-    print(HydraConfig.instance().hydra.job.name)
+    hc = HydraConfig.instance()
+    assert hc.hydra is not None
+    print(hc.hydra.job.name)
 
 
 if __name__ == "__main__":
