+++ hydra/core/hydra_config.py
@@ -1,6 +1,6 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
 import copy
-from typing import Any
+from typing import Any, Optional
 
 from omegaconf import DictConfig, OmegaConf
 
@@ -9,13 +9,13 @@ from hydra.core.singleton import Singleton
 
 
 class HydraConfig(metaclass=Singleton):
-    hydra: HydraConf
+    hydra: Optional[HydraConf]
 
     def __init__(self) -> None:
-        ret = OmegaConf.structured(HydraConf)
-        self.hydra = ret
+        self.hydra = None
 
     def set_config(self, cfg: DictConfig) -> None:
+        assert cfg is not None
         self.hydra = copy.deepcopy(cfg.hydra)
         OmegaConf.set_readonly(self.hydra, True)  # type: ignore
