+++ news/496.bugfix
@@ -1,0 +1,1 @@
+to_absolute_dir(path) now converts relative path to be relative to os.cwd() when used outside of Hydra
