+++ tests/test_config_repository.py
@@ -1,5 +1,5 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
-from typing import List, Optional
+from typing import Any, List, Optional
 
 import pytest
 
@@ -9,6 +9,7 @@ from hydra._internal.core_plugins.file_config_source import FileConfigSource
 from hydra._internal.core_plugins.package_config_source import PackageConfigSource
 from hydra._internal.core_plugins.structured_config_source import StructuredConfigSource
 from hydra.core.object_type import ObjectType
+from hydra.core.plugins import Plugins
 from hydra.test_utils.config_source_common_tests import ConfigSourceTestSuite
 from hydra.test_utils.test_utils import chdir_hydra_root
 
@@ -44,7 +45,8 @@ def create_config_search_path(path: str) -> ConfigSearchPathImpl:
     ],
 )
 class TestConfigRepository:
-    def test_config_repository_load(self, path: str) -> None:
+    def test_config_repository_load(self, restore_singletons: Any, path: str) -> None:
+        Plugins.instance()  # initializes
         config_search_path = create_config_search_path(path)
         repo = ConfigRepository(config_search_path=config_search_path)
         ret = repo.load_config(config_path="dataset/imagenet.yaml")
@@ -54,7 +56,8 @@ class TestConfigRepository:
         }
         assert repo.load_config(config_path="not_found.yaml") is None
 
-    def test_config_repository_exists(self, path: str) -> None:
+    def test_config_repository_exists(self, restore_singletons: Any, path: str) -> None:
+        Plugins.instance()  # initializes
         config_search_path = create_config_search_path(path)
         repo = ConfigRepository(config_search_path=config_search_path)
         assert repo.exists("dataset/imagenet.yaml")
@@ -76,11 +79,13 @@ class TestConfigRepository:
     )
     def test_config_repository_list(
         self,
+        restore_singletons: Any,
         path: str,
         config_path: str,
         results_filter: Optional[ObjectType],
         expected: List[str],
     ) -> None:
+        Plugins.instance()  # initializes
         config_search_path = create_config_search_path(path)
         repo = ConfigRepository(config_search_path=config_search_path)
         ret = repo.get_group_options(
