+++ tests/test_hydra.py
@@ -23,7 +23,10 @@ chdir_hydra_root()
     "calling_file, calling_module", [(".", None), (None, ".")]
 )
 def test_missing_conf_dir(
-    task_runner: TTaskRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    task_runner: TTaskRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
     with pytest.raises(MissingConfigException):
         with task_runner(
@@ -43,7 +46,10 @@ def test_missing_conf_dir(
     ],
 )
 def test_missing_conf_file(
-    task_runner: TTaskRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    task_runner: TTaskRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
     with pytest.raises(MissingConfigException):
         with task_runner(
@@ -63,7 +69,10 @@ def test_missing_conf_file(
     ],
 )
 def test_app_without_config___no_overrides(
-    task_runner: TTaskRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    task_runner: TTaskRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
     with task_runner(
         calling_file=calling_file,
@@ -83,7 +92,10 @@ def test_app_without_config___no_overrides(
     ],
 )
 def test_hydra_main_rerun(
-    task_runner: TTaskRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    task_runner: TTaskRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
     with task_runner(
         calling_file=calling_file,
@@ -103,7 +115,10 @@ def test_hydra_main_rerun(
     ],
 )
 def test_app_without_config__with_overrides(
-    task_runner: TTaskRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    task_runner: TTaskRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
     with task_runner(
         calling_file=calling_file,
@@ -126,7 +141,10 @@ def test_app_without_config__with_overrides(
     ],
 )
 def test_app_with_config_file__no_overrides(
-    task_runner: TTaskRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    task_runner: TTaskRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
 
     task = task_runner(
@@ -151,7 +169,10 @@ def test_app_with_config_file__no_overrides(
     ],
 )
 def test_app_with_config_path_backward_compatibility(
-    task_runner: TTaskRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    task_runner: TTaskRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
     task = task_runner(
         calling_file=calling_file,
@@ -176,7 +197,10 @@ def test_app_with_config_path_backward_compatibility(
     ],
 )
 def test_app_with_config_file__with_overide(
-    task_runner: TTaskRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    task_runner: TTaskRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
     with task_runner(
         calling_file=calling_file,
@@ -199,7 +223,10 @@ def test_app_with_config_file__with_overide(
     ],
 )
 def test_app_with_split_config(
-    task_runner: TTaskRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    task_runner: TTaskRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
     with task_runner(
         calling_file=calling_file,
@@ -222,7 +249,10 @@ def test_app_with_split_config(
     ],
 )
 def test_app_with_config_groups__override_dataset__wrong(
-    task_runner: TTaskRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    task_runner: TTaskRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
     with pytest.raises(MissingConfigException) as ex:
         with task_runner(
@@ -244,7 +274,10 @@ def test_app_with_config_groups__override_dataset__wrong(
     ],
 )
 def test_app_with_config_groups__override_all_configs(
-    task_runner: TTaskRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    task_runner: TTaskRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
     with task_runner(
         calling_file=calling_file,
@@ -267,7 +300,10 @@ def test_app_with_config_groups__override_all_configs(
     ],
 )
 def test_app_with_sweep_cfg__override_to_basic_launcher(
-    task_runner: TTaskRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    task_runner: TTaskRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
     with task_runner(
         calling_file=calling_file,
@@ -325,6 +361,7 @@ def test_cfg(tmpdir: Path, flag: str, expected_keys: List[str]) -> None:
 )
 @pytest.mark.parametrize("overrides", [["free_group=opt1,opt2"]])  # type: ignore
 def test_multirun_with_free_override(
+    restore_singletons: Any,
     sweep_runner: TSweepRunner,
     calling_file: str,
     calling_module: str,
@@ -355,7 +392,10 @@ def test_multirun_with_free_override(
     ],
 )
 def test_sweep_complex_defaults(
-    sweep_runner: TSweepRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    sweep_runner: TSweepRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
     with sweep_runner(
         calling_file=calling_file,
@@ -464,7 +504,10 @@ def test_help(
     ],
 )
 def test_interpolating_dir_hydra_to_app(
-    task_runner: TTaskRunner, calling_file: str, calling_module: str
+    restore_singletons: Any,
+    task_runner: TTaskRunner,
+    calling_file: str,
+    calling_module: str,
 ) -> None:
     basedir = "foo"
     with task_runner(
